+++ usr/src/usr.bin/f77/libF77/r_cosh.c
@@ -3,10 +3,10 @@
  * All rights reserved.  The Berkeley software License Agreement
  * specifies the terms and conditions for redistribution.
  *
- *	@(#)r_cosh.c	5.1	%G%
+ *	@(#)r_cosh.c	5.2	%G%
  */
 
-double r_cosh(x)
+float r_cosh(x)
 float *x;
 {
 double cosh();
